\name{cantilever}
\docType{data}
\alias{cantilever}
\title{A function calculating the deviation of a cantilever beam.}
\description{
The limit-state function is defined in the standard space and isoprobabilistic transformation is used internally.
}
\usage{cantilever}
\format{The function can handle a vector or a matrix with column vectors.}
\references{
Gayton, N. and Bourinet, J.-M. and Lemaire, M.:\cr
\emph{CD2RS: a new statistical approach to the response surface method for reliability analysis.}\cr
Structural Safety 25 99-121, 2003.\cr
}
\keyword{datasets}