\name{oscillator_d6}
\docType{data}
\alias{oscillator_d6}
\title{A limit-state-function defined with a non-linear oscillator in dimension 6.}
\description{
  The limit-state function is defined in the standard space and isoprobabilistic transformation is used internally.
}
\usage{oscillator_d6}
\format{The function can handle a vector or a matrix with column vectors.}
\references{
  Echard, B and Gayton, N and Lemaire, M and Relun, N:\cr
  \emph{A combined Importance Sampling and Kriging reliability method for
    small failure probabilities with time-demanding numerical models}\cr
  Reliability Engineering \& System Safety 111 232-240, 2013.\cr
}
\keyword{datasets}