% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputeUpdateData.R
\name{precomputeUpdateData}
\alias{precomputeUpdateData}
\title{precomputeUpdateData}
\usage{
precomputeUpdateData(model, integration.points)
}
\arguments{
\item{model}{a object from \code{km}}

\item{integration.points}{the points onto which the updated variance will be computed}
}
\value{
A list containing the following elements
    \item{Kinv.c.olddata}{kriging weights for the integrations.points over krig@X}
    \item{Kinv.F}{The matrix product of the inverse covariance and F the matrix of
    the trend functions at model@X}
   \item{first.member}{}
}
\description{
precomputeUpdateData
}
