\name{rackwitz}
\docType{data}
\alias{rackwitz}
\title{A limit-state-function defined by Rackwitz}
\description{
The function is defined in the standard space and internal normal-lognormal transformation is done. Its definition with iid lognormal random variables is:
\deqn{d + a\sigma\sqrt{d} - \sum_{i=1}^d x_i}
Default values are: \eqn{a=1}, \code{mean=1} and \eqn{\sigma=0.2}.
}
\usage{rackwitz}
\format{The function can handle a vector or a matrix with column vectors.}
\references{
  Rackwitz, R:\cr
  \emph{Reliability analysis: a review and some perspectives}\cr
  Structural Safety, 23, 4, 365-395, 2001.\cr
}

\keyword{datasets}