\name{twodof}
\docType{data}
\alias{twodof}
\alias{oscillator_d8}
\title{A limit-state-function defined with a two degrees of freedom damped oscillator}
\description{
  The limit-state function is defined in the standard space and isoprobabilistic transformation is used internally.
  
  Parameters \code{mean_Fs} and \code{p} can be specified and default are 27.5 and 3 respectively.
}
\usage{twodof}
\format{The function can handle a vector or a matrix with column vectors.}
\references{
  Dubourg, V and Deheeger, F and Sudret, B:\cr
  \emph{Metamodel-based importance sampling for the simulation of rare events}\cr
  arXiv preprint arXiv:1104.3476, 2011.\cr
}
\keyword{datasets}