% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSd_old.R
\name{updateSd.old}
\alias{updateSd.old}
\title{UpdateSd.old}
\usage{
updateSd.old(X.new, newdata.oldsd, model, precalc.data, integration.points)
}
\arguments{
\item{X.new}{the d x N matrix containing the points added to the model for
the update of the kriging variance.}

\item{newdata.oldsd}{a vector containing the standard deviation
of the points to be added to the metamodel learning database.}

\item{model}{the current kriging model (a \code{km} object).}

\item{precalc.data}{precomputed data from KrigInv::precomputeUpdateData.}

\item{integration.points}{points where the updated sd is to be calculated.}
}
\description{
UpdateSd.old
}
