% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtoX.R
\name{XtoU}
\alias{XtoU}
\title{From X to standard space}
\usage{
XtoU(X, input.margin, L0)
}
\arguments{
\item{X}{the matrix d x n of the input points}

\item{input.margin}{A list containing one or more list defining the marginal distribution
functions of all random variables to be used}

\item{L0}{the lower matrix of the Cholesky decomposition of correlation matrix R0
(result of \code{\link{ModifCorrMatrix}})}
}
\description{
XtoU lets transform datapoint in the original space X to the standard Gaussian
space U with isoprobalisitc transformation.
}
\seealso{
\code{UtoX}
}
\author{
Clement WALTER \email{clement.walter@cea.fr}
}
